[Feature] Provide an API for die rolls
This patch implements functions which generate random numbers using
the kind of notation that is found in most table-top role-playing
games, e.g. Dungeons and Dragons.  The fundamental function is
`chance.rpg()`, which accepts strings with said notation and returns
an array of results.  For example:

    chance.rpg("3d6") = { 2, 6, 3 }

The patch also implements the following shortcut functions for
commonly used die rolls:

    - chance.d4()
    - chance.d6()
    - chance.d8()
    - chance.d10()
    - chance.d12()
    - chance.d20()
    - chance.d100()

These each roll that type of die once and return the result as a
number, instead of an array like `chance.rpg()` returns.