[Feature] Allow collecting the results of a function into an array
If we wanted to create an array of ten random integers we could write
the following:

    local integers = {}
    for i = 1, 10 do
        table.insert(integers, chance.integer())
    end

This patch implements the function `chance.n()`, allowing us to
rewrite the above as simply this:

    local integers = chance.n(chance.integer, 10)

In other words, it allows us to execute a `chance` function N-number
of times while collecting the results.  This simplifies a common
use-case and pattern in code.

Finally, it is worth noting that `chance.n()` is not restricted to use
with other `chance` functions from the library.  The first argument to
`chance.n()` can be almost any function.  However, this behavior is
undocumented because it may not be supported in the future and thus no
code should rely on it.  The only reasoning for mentioning it here is
to act as a reminder that future functions in the API may be most
easily expressed in terms of the `chance.n()` function.