[Feature] Support randomly generating a gender
The topic of gender is an often debated one in modern society, and so
a function which randomly generates a gender is forced to tread
carefully with its output domain.  To that end, the `chance.gender()`
function will randomly select a string from the `chance.genders`
table, which developers can modify for their applications as needed.
After much thought I decided to go with the default values of "Male",
"Female", and "Third", preferring cultural identifications of gender
to medical classifications.  But even there the choice of default
values is not straight-forward.  Ultimately the best approach felt to
define a small default set of possible values and then make the values
easy to customize.  Developers who want to restrict the choices to
just "Male" or "Female" have to use the more verbose

    chance.gender { binary = true }

form of the function.  This design choice merely reflects the fact
that `chance.genders`, by default, has more than two values, and is
not intended to make any kind of "statement" about the role of gender
in modern cultures and societies.

See-Also: https://en.m.wikipedia.org/wiki/Third_gender