[Feature] Implement support for named data sets
This patch introduces the concept of "data sets" to the Core API.  A
data set is a predefined source of data which functions may choose
from at random.  The `chance.dataSets` table contains all of the data
sets, mapping both strings (as names of sets) or functions
(representing the function that uses the set) to the random data.
That data can be either a table containing any types of values, or it
can be a single function, in which case the function acts as a
generator of data, i.e. the library will invoke it whenever data is
requested for that set.

The API provides three functions for working with data sets:

1. `chance.set()` This function creates data sets, associating a set
with either a string, acting as the name of the set, or with a
function, indicating that the function relies on that data set.

2. `chance.appendSet()` This function adds data to an existing set.

3. `chance.fromSet()` This function randomly selects data from a given
set, thus acting as the foundation for any utilities built atop the
data set mechanism.

The idea moving forward is to rewrite many of the existing functions
to use data sets, as that will allow developers to redefine or modify
the data as needed.  For example, an application could redefine the
data set for `chance.month()` so that it selects names from a
different language.  However, this patch does not rewrite any of the
existing API in terms of data sets.  It only paves the way.