[Refactor] Remove the ability to use functions as data set keys
In immediate retrospect this functionality was not as useful as
originally predicted.  Attempting to redefine `chance.month()` in
terms of data sets exposed the problem:

    chance.set(chance.month, { ... })

    function chance.month()
        return chance.fromSet(chance.month)
    end

The function name `chance.month` cannot be used as a key within
`chance.month()` itself because the function is not defined at that
point, i.e. the function is trying to refer to itself before it's even
completely defined.

This patch redefines the data set API so that sets only use strings
for names instead of also accepting functions.  However, I am not
simply reverting or ammending the previous commit because in the
future I may want to reintroduce that functionality.  But for the time
being it would not serve a useful purpose.