[Refactor] Rename is_within_range() to simply within_range()
Current uses of the assertion look like this:

    assert.is_within_range(...)

However, Busted allows us to use `is` (or not) inherently in an
assertion, meaning we can also write the above like so:

    assert.is.is_within_range(...)

This example demonstrates the problem with the current name, as we
repeat `is`, which is not necessary.  This patch renames the custom
assertion function in order to improve readability in the test suite.