[Test] Provide an assertion for testing arrays for uniqueness
The two unit tests for `chance.unique()` have the same, copy-paste
style code for determining if the results of the function produces an
array of unique values.  This patch implements a new `unique_array()`
assertion which performs that same test but allows us to reduce the
amount of code in the unit tests themselves, allowing us to reduce
both tests for `chance.unique()` into simple one-liners.