[Test] Seed the random generator differently for tests
This improves the quality of the test suite since it makes the results
of the `chance` functions less predictable.  Or to put it another way,
it allows us to ensure that tests pass for randomized results instead
of basing the tests on a seed where the functions always return the
same value.  This helps a lot when we run the entire test suite
multiple times, as we do whenever we run Tup to build the project.