[Fix] Return correct values from chance.day() when using flags
If we use one of the optional flags for `chance.day()`, i.e. we call
the function in either of these forms

    chance.day { weekdays = false }
    chance.day { weekends = false }

then there is a chance the function will not return a correct value.
This happens because of the "clever" table manipulation we do inside
of `chance.day()` in order to restrict the possible return values.
This patch corrects the problem by restructuring the `days` data set
so that it has specific subsets for `"weekdays"` and `"weekends"`,
which allows us to rewrite `chance.day()` in such a way that it
selects from the proper subtables without the need for any
manipulation of said tables.