[Feature] Allow chance.day() to accept boolean true flags
Currently `chance.day()` allows selecting weekdays or weekends by
using boolean false with the optional flags like so:

    chance.day { weekdays = false }

That code returns a weekend.  This patch modifies the functions so
that those flags accept boolean true values, meaning we can rewrite
the above code like so:

    chance.day { weekends = true }

This can improve readability and intent, and at the very least makes
the API more consistent, as it was strange the function accepted only
false values.