[Fix] Correct the return value and error messages of like_pattern()
This patch first changes the `like_pattern()` assertion so that it
returns a boolean, making it more consistent with our other custom
assertions as they all do the same, and this brings the function more
in line with the luassert library in general.

Second, the patch corrects the error messages for the assertion.
Currently they show the pattern and string-to-test in the incorrect
order.  The patch rewords the messages so the assertion's arguments
appear in the correct order.