[Feature] Allow the creation of greyscale colors via chance.color()
The feature in this patch is conceptually simple: the new `greyscale`
flag causes `chance.color()` to return a greyscale color.  However,
the implementation was cause for an entire rewrite of the function.
The new implementation focuses more on generating random values for
red, green, and blue separately, and then combines them into the
proper type of string as needed.  This approach makes it easier to
implement the `greyscale` flag because by having separate variables
for red, green, and blue, we can create a greyscale color by simply
making green and blue equal to red.