[Feature] Add a feature to generate a Poker deck
This patch implements the `chance.poker.deck()` function.  It creates
a standard deck of 52 cards by default, but can optionally include the
Joker for a total of 53 cards.  With regard to the Joker there is a
disparity between this function and `chance.poker.card()`, which may
return the Joker by default.  The reasoning is that if users want an
individual card then it is more covenient to include the Joker in the
possibilities, under the assumption that users will more often want
*any* type of card whenever they generate them individuall.  However,
when users want a complete deck at once they will more often *not*
want the Joker since popular forms of the game do not use that card.