[Feature] Add a function to randomly generate Poker hands
This patch implements a function that generates a random hand of cards
for Poker, building upon the existing functionality for creating
individual cards and decks.  The function makes hands of five cards by
default under the assumption that five cards is the most common
use-case, e.g. it's the number of cards used for hands in Draw,
Lowball, Duece-to-Seven, and the community cards in Texas Hold'em.